LayoutTests/imported/w3c/web-platform-tests/content-security-policy/gen/top.http-rp/worker-src-none/sharedworker-module.http-expected.txt


Harness Error (FAIL), message = Unhandled rejection: null is not an object (evaluating 'resolveObject.addEventListener')

FAIL Content Security Policy: Expects blocked for sharedworker-module to same-http origin and keep-origin redirection from http context. promise_test: Unhandled rejection with value: object "TypeError: null is not an object (evaluating 'worker.port.start')"
FAIL Content Security Policy: Expects blocked for sharedworker-module to same-http origin and keep-origin redirection from http context.: securitypolicyviolation assert_equals: One violation event should be fired expected 1 but got 0
FAIL Content Security Policy: Expects blocked for sharedworker-module to same-http origin and no-redirect redirection from http context. promise_test: Unhandled rejection with value: object "TypeError: null is not an object (evaluating 'worker.port.start')"
FAIL Content Security Policy: Expects blocked for sharedworker-module to same-http origin and no-redirect redirection from http context.: securitypolicyviolation assert_equals: One violation event should be fired expected 1 but got 0
